fix: size of input mask for count_least_significant_bits

+++ include/cuco/detail/utils.cuh
@@ -48,10 +48,10 @@ namespace cuco {
 namespace detail {
 
 /**
- * @brief For the `n` least significant bits in the given unsigned 32-bit integer `x`,
+ * @brief For the `n` least significant bits in the given unsigned 64-bit integer `x`,
  * returns the number of set bits.
  */
-__device__ __forceinline__ int32_t count_least_significant_bits(uint32_t x, int32_t n)
+__device__ __forceinline__ int32_t count_least_significant_bits(uint64_t x, int32_t n)
 {
   return __popcll(x & (1UL << n) - 1UL);
 }
